Component tiles timer
